requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
matplotlib>=3.7.0
google-generativeai>=0.3.0
